+++ mne/source_space.py
@@ -21,7 +21,7 @@ from .io.write import (start_block, end_block, write_int,
                        write_float_sparse_rcs, write_string,
                        write_float_matrix, write_int_matrix,
                        write_coord_trans, start_file, end_file, write_id)
-from .bem import read_bem_surfaces
+from .bem import read_bem_surfaces, ConductorModel
 from .surface import (read_surface, _create_surf_spacing, _get_ico_surface,
                       _tessellate_sphere_surf, _get_surf_neighbors,
                       _normalize_vectors, _get_solids, _triangle_neighbors,
@@ -1458,10 +1458,11 @@ def setup_volume_source_space(subject=None, pos=5.0, mri=None,
         interpolation matrix over. Source estimates obtained in the
         volume source space can then be morphed onto the MRI volume
         using this interpolator. If pos is a dict, this can be None.
-    sphere : array_like (length 4)
+    sphere : ndarray, shape (4,) | ConductorModel
         Define spherical source space bounds using origin and radius given
-        by (ox, oy, oz, rad) in mm. Only used if `bem` and `surface` are
-        both None.
+        by (ox, oy, oz, rad) in mm. Only used if ``bem`` and ``surface``
+        are both None. Can also be a spherical ConductorModel, which will
+        use the origin and radius.
     bem : str | None
         Define source space bounds using a BEM file (specifically the inner
         skull surface).
@@ -1538,9 +1539,17 @@ def setup_volume_source_space(subject=None, pos=5.0, mri=None,
                                  'check  freesurfer lookup table.'
                                  % (label, mri))
 
-    sphere = np.asarray(sphere)
+    if isinstance(sphere, ConductorModel):
+        if not sphere['is_sphere'] or len(sphere['layers']) == 0:
+            raise ValueError('sphere, if a ConductorModel, must be spherical '
+                             'with multiple layers, not a BEM or single-layer '
+                             'sphere (got %s)' % (sphere,))
+        sphere = tuple(1000 * sphere['r0']) + (1000 *
+                                               sphere['layers'][0]['rad'],)
+    sphere = np.asarray(sphere, dtype=float)
     if sphere.size != 4:
-        raise ValueError('"sphere" must be array_like with 4 elements')
+        raise ValueError('"sphere" must be array_like with 4 elements, got: %s'
+                         % (sphere,))
 
     # triage bounding argument
     if bem is not None: